==> rscripts/0_Obtain_SpatialData.Rmd <==
---
title: "DataPreparation_100m"
author: "Aimara Planillo"
date: "2023-10-30"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Prepare environmental variables

```{r prepare workspace}
source("./rscripts/source_pkgs.R")

wrkdir <- getwd()
geoproc_wd <- paste0(wrkdir, "/output/geo_proc")
```


### load german borders
```{r load german borders}
## german borders
germany <- st_read("./data/geo_raw/germany_3035.gpkg")
plot(st_geometry(germany))

## get only the outside border
german_border <- st_union(germany)
plot(st_geometry(german_border))
```

# load raster mosaics and save environmental layers for Germany


## imperviousness (copernicus), year 2018, 10 m resolution, epsg; 3035

```{r load copernicus data}
## imperviousness 2018 at 10m res
imp <- rast("D:/Nextcloud/GeoData/data-raw/germany/Imperviousness_density_germany_2018_10m_03035_XXX_tif/IMD_2018_010m_E40N28_03035_v020.tif")
## this path changes with the computer!!

## look at data
plot(imp, add = TRUE) 

## load all layers for Germany
imp_l <- lapply(
  list.files(
  path = "D:/Nextcloud/GeoData/data-raw/germany/Imperviousness_density_germany_2018_10m_03035_XXX_tif",
  pattern = ".tif$",
  recursive = FALSE,
  full.names = TRUE), rast)

## put together in one raster
imp_ger <- do.call(terra::mosaic, imp_l)

## cut to German borders
imp_ger_c <- mask(imp_ger, vect(german_border))

plot(imp_ger_c)

## set NA values (255) to NA 
NAflag(imp_ger_c) <- 255

## correct name of raster
names(imp_ger_c) <- "imperv_perc"

## save raster for later use
# writeRaster(imp_ger_c, paste0(geoproc_wd, "/imperviousness_germany_2018_10m_3035.tif"), overwrite = TRUE)
# imp_ger_c <- rast( paste0(geoproc_wd, "/imperviousness_germany_2018_10m_3035.tif"))
```


## tree cover density (copernicus), year 2018, 10 m resolution, epsg; 3035

```{r}
tcd_l <- lapply(
  list.files(
  path = "D:/Nextcloud/GeoData/data-raw/europe/tree_cover_density_2018_10m_3035/TCD_2018_010m_de_03035_v020",
  pattern = ".tif$",
  recursive = TRUE,
  full.names = TRUE), rast)

tcd_ger <- do.call(terra::mosaic, tcd_l)

tcd_ger_c <- mask(tcd_ger, vect(german_border))

NAflag(tcd_ger_c) <- 255
names(tcd_ger_c) <- "tcd"
plot(tcd_ger_c)

# writeRaster(tcd_ger_c, paste0(geoproc_wd, "/tcd_germany_2018_10m_3035.tif"), overwrite = TRUE)
# tcd_ger_c <- rast( paste0(geoproc_wd, "/tcd_germany_2018_10m_3035.tif"))
```


## corine landcover (copernicus), year 2018, polygon file, epsg; 25832 to epsg:3035

```{r}
corine_2018_l <- lapply(
  list.files(
  path = "D:/Nextcloud/GeoData/data-raw/germany/Corine_Landcover_2018_25832_shp/300000765_1330.lbm-de2018/lbm-de2018",
  pattern = ".shp$",
  recursive = TRUE,
  full.names = TRUE), st_read)

corine_2018 <- do.call(rbind, corine_2018_l)

## correct crs
corine_2018_3035 <- st_transform(corine_2018, crs(tcd_ger_c))

corine_2018_3035_sub <- corine_2018_3035[,"CLC18"]

## rasterize using tcd as template
corine_2018_ras <- rasterize(corine_2018_3035_sub,
                             tcd_ger_c,
                            field = "CLC18", 
                            fun = "max")

corine_2018_ras_c <- mask(corine_2018_ras, german_border)
values(corine_2018_ras_c)
names(corine_2018_ras_c) <- "clc2018"
plot(corine_2018_ras_c)

# writeRaster(corine_2018_ras_c, paste0(geoproc_wd, "/clc_germany_2018_10m_3035.tif"), overwrite = TRUE)
# corine_2018_ras_c <- rast(paste0(geoproc_wd, "/clc_germany_2018_10m_3035.tif"))
```


# distance to roads 10m

Distance to asphalted roads

This data was obtained from https://gdz.bkg.bund.de
https://gdz.bkg.bund.de/index.php/default/digitale-geodaten/digitale-landschaftsmodelle.html?___store=default
Digitales Basis-Landschaftsmodell (Ebenen) (Basis-DLM) 

**This source costs money and it is not publicly available**
PROBLEMS WITH PUBLICATIONS?? -- CHECK FOR OPEN ACCESS SOURCES!!

I use XXXX in the name of the file instead of the year, as it is unknown for the moment.

```{r}
road_list <- lapply(list.files(
  path = paste("D:", "Nextcloud", "GeoData", "data-raw", "germany", "digital_base_dlm_germany_2022_25832_BKG_shp", "300001721_2890.basis-dlm-aaa_ebenen", "basis-dlm-aaa_ebenen", sep = "/"), 
  full.names = TRUE,
  recursive = TRUE, 
  pattern = "ver01_l.shp$"
), st_read)

road_ger <- do.call(rbind, road_list)

road_ger_3035 <- road_ger %>% st_transform(crs(imp_ger_c)) %>% mutate(rast = 1) %>% dplyr::select(rast)
road_rast <- rasterize(road_ger_3035, imp_ger_c, field = "rast")

road_dist <- terra::distance(road_rast)
road_dist_c <- crop(road_dist, german_border)
road_dist_m <- mask(road_dist, german_border)
plot(road_dist_m)

# not: In the south-east are no roads. 
# writeRaster(road_dist_m, paste0(geoproc_wd, "/distance_to_roads_germany_XXXX_10m_3035.tif"), overwrite = TRUE)
# road_dist_m <- rast(paste0(geoproc_wd, "/distance_to_roads_germany_XXXX_10m_3035.tif"))
```


# distance to paths 10m

Distance to hiking / cycling paths

SAME ORIGIN AS LAYER ABOVE!!

```{r}
# USE imperviousness map as template
imp_ger_c 

path_list <- lapply(list.files(
  path = paste("D:", "Nextcloud", "GeoData", "data-raw", "germany", "digital_base_dlm_germany_2022_25832_BKG_shp", "300001721_2890.basis-dlm-aaa_ebenen", "basis-dlm-aaa_ebenen", sep = "/"), 
  full.names = TRUE,
  recursive = TRUE, 
  pattern = "ver02_l.shp$"
), st_read)

path_ger <- do.call(rbind, path_list)
path_ger_3035 <- path_ger %>% st_transform(crs(imp_ger_c)) %>% mutate(rast = 1)  %>% dplyr::select(rast)

path_rast <- rasterize(path_ger_3035, imp_ger_c, field = "rast")

path_dist <- terra::distance(path_rast)
path_dist_c <- crop(path_dist, german_border)
path_dist_m <- mask(path_dist, german_border)

plot(path_dist_m)
# writeRaster(path_dist_m,  paste0(geoproc_wd, "/distance_to_paths_germany_XXXX_10m_3035.tif"), overwrite = TRUE)
# path_dist_m <- rast(paste0(geoproc_wd, "/distance_to_paths_germany_XXXX_10m_3035.tif"))
```



# Precipitation: dwd https://opendata.dwd.de

Resolution: 1 km, crs = 31467

Monthly total precipitation (l/m2)

Precipitation data, obtained from the Deutscher Wetterdients
https://opendata.dwd.de
More info in:
https://www.dwd.de/EN/ourservices/opendata/opendata.html

DWD should be cited as the source of data:
https://www.dwd.de/EN/service/copyright/templates_dwd_as_source.html?nn=495490

```{r}
### this is working
# f <- basename(url)
# download.file(url, f, mode="wb")
# y <- rast(f)
# 
# plot(y)

### monthly precipitation data
url <- 'https://opendata.dwd.de/climate_environment/CDC/grids_germany/monthly/precipitation/08_Aug/'

url_names <- readLines(url)
#url_names[[5]]

list_url <- lapply(5:(length(url_names)-2), function(iter){
stringi::stri_split(url_names[[iter]], regex = "\\\"")[[1]][2]})

url_df <- as.data.frame(do.call("rbind", list_url))
names(url_df) <- "url"

my_years <- c("2015", "2016", "2017", "2018")
my_urls <- url_df %>% 
  filter(stringr::str_detect(url, pattern = paste(my_years, collapse = "|")))
my_urls


glue::glue(url, my_urls[1,])
## download data
for(i in 1:nrow(my_urls)){
  download.file(glue::glue(url, my_urls[i,]), 
                destfile = paste0( "./data/geo_raw/prec/", my_urls[i,]), mode="wb")
  R.utils::gunzip(paste0("./data/geo_raw/prec/", my_urls[i,]), remove=FALSE)

}

## load downloaded files 
my_files <- list.files("./data/geo_raw/prec/", pattern = paste(my_years, collapse = "|"), full.names = TRUE)
my_files <- my_files[!grepl(".gz", my_files)]

prec_rasters <- rast(my_files)
names(prec_rasters) <- paste0("prec_aug_", my_years)  
crs(prec_rasters) <- "epsg:31467"

plot(prec_rasters$prec_aug_2015)
plot(prec_rasters$prec_aug_2016)
plot(prec_rasters$prec_aug_2017)
plot(prec_rasters$prec_aug_2018)

res(prec_rasters)
```

# dwd soil moisture

SAME SOURCE AS ABOVE

Description:

Soil moisture is shown as a percentage of effective field capacity ( %nFK) for winter grain on lighter soils and for sugar beet on heavier soils, in each case for every month, season and year.

The effective field capacity (nFK) designates the volume of water already in the soil and which is available for the plants growing on it. Below a value of 50%nFK, plants must be assumed to suffer water stress. Values of over 100%nFK mean that soils are oversaturated. The calculations are made on a monthly basis for the entire calendar year even if no plants at all are growing on the monitored fields at certain times.

Values are based on calculations made useing the tried and tested SVAT model AMBAV. The AMBAV model uses the classic Penman-Monteith equation (Click here for more information). This model is routinely used in agrometeorological consulting work (e.g. irrigation advice). AMBAV results for the future are currently only available for 16 of the regional climate models (14 numerical and 2 statistical models). The input data needed for some calculations is unavailable for the remaining five models. The depiction of the trend for the average in Germany includes the results from 14 numerical models and the results from 2 statistical models.

```{r}
url <- "https://opendata.dwd.de/climate_environment/CDC/grids_germany/multi_annual/soil_moist/grids_germany_multi_annual_soil_moist_1991-2020_12.asc.gz"
f <- basename(url)
download.file(url, paste0("./data/geo_raw/soil_moisture/", f), mode="wb")
R.utils::gunzip(paste0("./data/geo_raw/soil_moisture/", f), remove=FALSE)

f_2 <-  gsub(".gz", "", f)
ras_moist <- rast(paste0("./data/geo_raw/soil_moisture/", f_2))

crs(ras_moist) <- st_crs(31467)$wkt

plot(ras_moist)
plet(ras_moist, tiles = "Streets")
```


# thünen soil ph

https://atlas.thuenen.de/layers/pH_map_30_100:geonode:pH_map_30_100

Data with about 1 km resolution
### looks like data has some gaps

```{r}
soil_ph <- rast("./data/geo_raw/thuenen_soil_ph_2021.tiff")
soil_ph
# plet(soil_ph, tiles = "Streets")

# transform to epsg 3035
soil_ph_3035 <- project(soil_ph, "epsg:3035")
res(soil_ph)
res(soil_ph_3035)
plet(soil_ph_3035, tiles = "Streets")
```


# thuenen livestock data

Polygon data, epsg 4326

Jahr 2020, 
Landwirtschaftliche Fläche (Code UAAR) in 1000 ha, ANTO: 

Großvieheinheiten in 1000 GVE; 
      CATO: Rinder; 
      PITO: Schweine; 
      POTO: Geflügel; 
      RETO: Andere Tiere; 
      DCOW: Milchvieh; 
      SCOW: Mutterkühe; 
      BULL: Rindermast; 
      CALV: Kälber; 
      HEIT: Färsen; 
      SOWS: Ferkel und Zuchtsauen; 
      PIGF: Mastschweine; 
      HENS: Hennen; 
      POUF: Mastgeflügel; 
      SHGM: Schafe und Ziegen; 
      OANI: Andere Tiere (Pferde)
       alle Tiere ebenfalls in 1000 GVE, Geflügel in 1000000 GVE

translation: 

Livestock units in 1000 GVE; 
**This value represents the amount of livestock measured in "1000 livestock units" (1000000 units for poultry)**
Cannot be transformed into real animal numbers, but should be comparable across regions.
To obtain a density, we should divide this number by the area.
      ANTO: all livestock
      CATO: Cattle; the sum of all cattle related subclasses
      PITO: Pigs; the sum of all pig related subclasses
      POTO: Poultry; the sum of all poultry related subclasses
      RETO: other animals; 
      DCOW: Dairy cattle; 
      SCOW: Mother cows; 
      BULL: Beef fattening; 
      CALV: calf; 
      HEIT: Heifers; 
      SOWS: Piglets und Breeding sows; 
      PIGF: Fattening pigs; 
      HENS: Hens; 
      POUF: Poultry for fattening; 
      SHGM: Sheeps und Goats; 
      OANI: other animals (Horses)  
     all animals in 1000 GVE, Poultry in 1000000 GVE


Definition german: Grundlage ist das Lebendgewicht. Eine Großvieheinheit entspricht etwa dem Gewicht eines ausgewachsenen 500 Kilogramm schweren Rindes.

Beispiele:

    Kalb = 0 4 GV
    junge Kuh = 0 6 GV
    Eber = 0 3 GV
    Mastschwein = 0 12 GV
    Ferkel = 0 01 GV
    Pferd = 1 GV
    Schaf = 0 1 GV
    Legehenne = 1 durch 320 GV.


https://www.ima-agrar.de/wissen/agrilexikon/grossvieheinheit-gv

Definition english: The basis is the live weight. A livestock unit is roughly equivalent to the weight of a full-grown 500-kilogram cow. 

Example:

    calf = 0 4 GV
    young cow Kuh = 0 6 GV
    wild boar = 0 3 GV
    pig = 0 12 GV
    pglet = 0 01 GV
    horse = 1 GV
    sheep = 0 1 GV
    laying hen = 1 durch 320 GV.



```{r}
livestock <- st_read("./data/geo_raw/thuenen_livestock/Gemeinde_Tiere_LU_2020_EXT_MWM_01_2023-09-18_09-37-30.shp") %>% 
  st_transform(3035)


## Merge with regional data (obtained from Destatis) 

## load regional data
munip <- readxl::read_xlsx("./data/geo_raw/thuenen_livestock/31122010_Auszug_GV.xlsx",
                             sheet = 2,
                             skip = 5,
                             col_names = c("Satzart",
                                           "txt_kenn",
                                           "RS_Land",
                                           "RS_RB",
                                           "RS_Kreis",
                                           "RS_VB",
                                           "RS_Gem",
                                           "gem_name",
                                           "flaeche_km2",
                                           "Bev_insgesamt",
                                           "Bev_maennl",
                                           "Bev_weibl",
                                           "Bev_je_km2",
                                           "PLZ",
                                           "lon",
                                           "lat")) %>%  
  mutate(id = 1:n()) %>% 
  drop_na(c(RS_Land, RS_RB, RS_Kreis, RS_Gem)) %>% 
  group_by(id) %>%  
  mutate(gem_char = paste0(RS_Land, RS_RB, RS_Kreis, RS_Gem)) %>% 
  ungroup()

## merge livestock and regions
livestock_fixed <- livestock %>%  
  select(-gem_name) %>% 
  left_join(munip %>% 
              select(gem_char, gem_name), 
            by = "gem_char")
   

## check that the projection is the correct one
livestock_fixed <- livestock_fixed %>% 
  st_transform(3035)

leaflet::leaflet(livestock_fixed %>% 
                   st_transform(crs = 4326)) %>% 
  leaflet::addPolygons() %>% 
  leaflet::addTiles()

# st_write(livestock_fixed, paste0(geoproc_wd, "/livestock_germany_2020_vector_3035.gpkg"))
# livestock_fixed <- st_read(paste0(geoproc_wd, "/livestock_germany_2020_vector_3035.gpkg"))


livestock_fixed %>% 
  filter(gem_name == "Berlin, Stadt")

livestock %>% 
  filter(gem_name == "Berlin")
```


############################################
# Environmental layers at 100 m resolution #
############################################

## aggregate rasters

We make the rasters 100 m resolution to diectly extract the environmental values at this resolution for the mice trapping points in the next step

## imperviousness

```{r imperv}
imp_ger_c
## make the NA zeros
NAflag(imp_ger_c) <- 255

imp_100m <- terra::aggregate(imp_ger_c, fact = 10, fun = "mean", na.rm = TRUE)
plot(imp_100m)
names(imp_100m) <- "imperv"
imp_100m

writeRaster(imp_100m, paste0(geoproc_wd, "/imperviousness_2018_100m_3035.tif"), overwrite = TRUE)
```

## tree cover density

```{r tcd}
tcd_ger_c
NAflag(tcd_ger_c) <- 255

tcd_100m <- terra::aggregate(tcd_ger_c, fact = 10, fun = "mean", na.rm = TRUE)
plot(tcd_100m)
names(tcd_100m) <- "tcd"

writeRaster(tcd_100m, paste0(geoproc_wd, "/tcd_2018_100m_3035.tif"), overwrite = TRUE)
```

## corine land cover

```{r clc}
corine_2018_ras_c

clc_100m <- aggregate(corine_2018_ras_c, fact = 10, fun = "modal", na.rm = TRUE)
plot(clc_100m)
plot(german_border, add = TRUE, col = "transparent", border = "red", lwd = 4)
names(clc_100m) <- "clc"

writeRaster(clc_100m, paste0(geoproc_wd, "/clc_2018_germany_100m_3035.tif"))
```

## swf raster 100 m
Copernicus: small wood features, provided as categorical raster at 100m res for 2018
```{r}

## load european data
swf_100m_tmp <- rast("./data/geo_raw/swf_2018/SWF_2018_100m_eu_03035.tif")
plot(swf_100m)

swf_100m_tmp <- project(swf_100m, st_crs(german_border)$wkt)

## crop to germany
swf_germany <- crop(swf_100m_tmp, german_border)

plot(swf_germany)
plot(german_border, add = TRUE, col = "transparent", border = "red", lwd = 4)

## to make numerical
rast_df <- cats(swf_germany) %>% 
  as.data.frame() %>% 
  mutate(swf_perc = Value)

swf_tmp <- categories(swf_germany, 
                          layer = 1, 
                          value = rast_df, 
                          active = 4) # column 4 (do not count the value column! It has to be numeric)

swf_num_tmp <- catalyze(swf_tmp, index = 4) # column 3 for new_value

# we select only the correct numerical column for the new raster
swf_num <- swf_num_tmp$swf_perc
NAflag(swf_num) <- 255

## resample to imperv raster to make them stack
swf_100m <- terra::resample(swf_num, imp_100m)
NAflag(swf_100) <- 255

swf_100m <- mask(swf_100m, vect(german_border))
plot(swf_100)

#  writeRaster(swf_100, paste0(geoproc_wd, "/swf_2018_germany_100m_3035.tif"), overwrite = TRUE)
# swf_100 <- rast(paste0(geoproc_wd, "/swf_2018_germany_100m_3035.tif"))
```


## distance to roads paths 100m

```{r}
road_d_100m <- terra::aggregate(road_dist_m, fact = 10, fun = "mean", na.rm = TRUE)
road_d_100m
names(road_d_100m) <- "dist_road"
plot(road_d_100m)

writeRaster(road_d_100m, paste0(geoproc_wd, "/distance_roads_germany_2018_100m_3035.tif"))
```

## distance to paths 100m

```{r}
path_d_100m <- terra::aggregate(path_dist_m, fact = 10, fun = "mean", na.rm = TRUE)
path_d_100m
names(path_d_100m) <- "dist_path"
plot(path_d_100m)

writeRaster(path_d_100m, paste0(geoproc_wd, "/distance_to_paths_germany_2018_100m_3035.tif"))
```


## precipitation
We create one for each year and resample it to 100m so it stacks with the other layers
```{r}
crs(prec_rasters)
crs(imp_100m)

prec_3035 <- project(prec_rasters, crs(imp_100m))

prec_2015_100m <- resample(prec_3035$prec_aug_2015, imp_100m)
prec_2016_100m <- resample(prec_3035$prec_aug_2016, imp_100m)
prec_2017_100m <- resample(prec_3035$prec_aug_2017, imp_100m)
prec_2018_100m <- resample(prec_3035$prec_aug_2018, imp_100m)

plot(prec_2016_100m)

```

## soil moisture
resampling to match extent and resolution of the other layers
```{r}
ras_moist
moist_3035 <- project(ras_moist, crs(imp_100m))
moist_100m <- resample(moist_3035, imp_100m)
plot(moist_100m)
```

## soil ph
resampling to match extent and resolution of the other layers
```{r}
soil_ph_3035

soil_ph_100m <- resample(soil_ph_3035, imp_100m)
plot(soil_ph_100m)
```

## rasterize livestock info
```{r}
## get areas
livestock_fixed$area_km <- st_area(livestock_fixed) %>% 
  units::set_units(km^2)

livestock_fixed$area_ha <- units::set_units(livestock_fixed$area, ha)
livestock_fixed

## Make spatvector
livestock_vct <- vect(livestock_fixed)

# raster template 100 m
raster100 <- rast(imp_100m)
# create a smaller one to assign each cell the value of the polygon overlapping the center adn them merge to the max overlap
n <- 10
rast10 <- disagg(raster100, n)


## municipalities
rast_mun <- rasterize(livestock_vct, rast10, field = "gem_char")
municipalities <- aggregate(rast_mun, n, "modal")
plot(municipalities)

# writeRaster(municipalities, paste0(geoproc_wd, "/germany_municipalities_100m_3035.tif"))
```

```{r livestock raw numbers}
## Cattle
rast_cattle <- rasterize(livestock_vct, rast10, field = "CATO")
cattle <- aggregate(rast_cattle, n, "modal")
plot(cattle)

## Pigs
rast_pigs <- rasterize(livestock_vct, rast10, field = "PITO")
pigs <- aggregate(rast_pigs, n, "modal")

### Poultry
rast_poultry <- rasterize(livestock_vct, rast10, field = "POTO")
poultry <- aggregate(rast_poultry, n, "modal")


writeRaster(cattle, paste0(geoproc_wd, "/cattle_germany_2020_100m_3035.tif"))
writeRaster(pigs, paste0(geoproc_wd, "/pigs_germany_2020_100m_3035.tif"))
writeRaster(poultry, paste0(geoproc_wd, "/poultry_germany_2020_100m_3035.tif"))
```

To get **densities**, we calculate the **livestock units / ha**. 
We do this because if we keep the original values (1000 livestock units), the numbers become really small. 
```{r livestock densities}
## make livestock a density by Km2 for cattle, pigs and poultry
livestock_fixed <- livestock_fixed %>% 
  mutate(cat_dens = CATO / as.numeric(area_ha) * 1000, ## we multiply for the value of the GVE (1000 GVE , 500kg cow) to avoid very low values
         pig_dens = PITO / as.numeric(area_ha) * 1000, ## we multiply for the value of the GVE (1000 GVE , 500kg cow) to avoid very low values
         pou_dens = POTO / as.numeric(area_ha) * 1000000) %>%  ## we multiply for the value of the GVE (1000000 GVE , 500kg cow) to avoid very low values
  dplyr::select(gem_char, gem_name, CATO, PITO, POTO, area_ha, cat_dens, pig_dens, pou_dens) 
  
livestock_dens
# The value here represents the livestock units / ha
  
livestock_vct <- vect(livestock_fixed)

## Cattle density
rast_catdens <- rasterize(livestock_vct, rast10, field = "cat_dens")
ext(rast_catdens)
catdens <- aggregate(rast_catdens, n, fun = "mean", na.rm = TRUE)
ext(catdens)
plot(catdens)

## Pigs
rast_pigdens <- rasterize(livestock_vct, rast10, field = "pig_dens")
pigdens <- aggregate(rast_pigdens, n, "mean")

### Poultry
rast_poudens <- rasterize(livestock_vct, rast10, field = "pou_dens")
poudens <- aggregate(rast_poudens, n, "mean")

writeRaster(catdens, paste0(geoproc_wd, "/cattle_GVE_densityha_germany_2020_100m_3035.tif"), overwrite=TRUE)
writeRaster(pigdens, paste0(geoproc_wd, "/pigs_GVE_densityha_germany_2020_100m_3035.tif"), overwrite=TRUE)
writeRaster(poudens, paste0(geoproc_wd, "/poultry_GVE_densityha_germany_2020_100m_3035.tif"), overwrite=TRUE)
```

## Put all environmental layers together
```{r}
env_cov_100m <- c(imp_100m, tcd_100m, clc_100m, swf_100m, 
                  road_d_100m, path_d_100m, 
                  prec_2015_100m, prec_2016_100m, prec_2017_100m, prec_2018_100m,
                  moist_100m, soil_ph_100m, 
                  cattle, pigs, poultry, catdens, pigdens, poudens, 
                  municipalities)
env_cov_100m

my_envrasters <- data.frame(names = names(env_cov_100m))

writeRaster(env_cov_100m, paste0(geoproc_wd, "/stack_env_covariates_germany_100m_3035.tif"), overwrite = TRUE)
write.csv(my_envrasters, paste0(geoproc_wd, "/stack_env_covariates_germany_100m_3035_names.csv"), row.names = FALSE)

ext(imp_100m)
ext(tcd_100m)
ext(clc_100m)
ext(swf_100m)
```
