==> rscripts/1_PrepareVariables.Rmd <==
---
title: "Variable_extraction"
author: "Aimara Planillo and Emanuel Heitlinger"
date: "2023-12-11"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Prepare variables for analysis

Using jS_ as a prefix for the variables selected in this project part (for this manuscript)

```{r prepare workspace}
source("rscripts/source_pkgs.R")
```


the data in object still has a data losing transformation (relative
abundances added up over different amplicons) in @otu_table. We
will replace this by the complete data  
```{r load data}
PS.TSS_filtered <- readRDS("./data/PS.TSS_filtered.rds")
```


#####################
# Prepare mice data #
#####################

```{r load data}
## N mice trapped per location
aggregateDensity <- unclass(sample_data(PS.TSS_filtered)) %>%
    ## need to remove class "sample_data" (phyloseq)
    as_tibble() %>%
    ## need to remove NA the same data 
    filter(!is.na(Body_Length)) %>% 
    ## add total mice trapped per locality and year  
    group_by(Locality, Year) %>%
    ## Number of mice trapped at one locality (combined long/lat at
    ## 4-digit resolution, available in "Locality"). We processed all
    ## mice for each locality. We create the variable per year to
    ## estimate density at one point in time. Hypothesis: as measure
    ## of host population density this indicates increase transmission
    ## between mice. More host-associated transmissible taxa will be
    ## enriched. Potential alternatives not considered now:
    ## collapsing localities at larger scale (more rounding),
    ## persistent density over years for those)
    summarize(jS_Ntrapped = n())  

## filter NAs and select variables
allSData <- unclass(sample_data(PS.TSS_filtered)) %>%
    ## need to remove class "sample_data" (phyloseq)
    as_tibble() %>%
    ## mice needing to be dropped because of NA values (might be
    ## worthwhile to imput this at some point, "Status" could be most
    filter(!is.na(Body_Length)) %>% 
    ## add here other variable that could need filtering if included
    ## SELECTION FOR NOW:
    ## Latitude and longitude: the basic spatial coordinates, decimal
    ## format, rounded to 4 digits
    mutate(jS_mouseID = Mouse_ID,
           jS_sex = Sex, 
           jS_lon = Longitude, 
           jS_lat = Latitude, 
           ## year: the year of sampling, hypothesis: closer years are more
           ## similar as a result of community persistence or similarity of
           ## environmental condition in closer years.
           jS_year = Year,
           ## BMI: the residuals of body weight on body
           ## length. Hypothesis: mice with worse body condition have
           ## disturbed microbiomes either as a cause or consequence
           ## of bad health. Might be worthwhile to consider plain
           ## weight or lenght as a proxy for age?! It'd be actually
           ## btter to use tail lenght for that, but not for now...
           jS_BMI = residuals(lm(Body_Weight~Body_Length, data = .)),
           ## HI, the hybrid index. Hypothesis: "disturbed"
           ## microbiomes (especially mycobiomes) in hybrids as
           ## observed previously.
           jS_HI = HI,
           ## Remark: All species interaction hypotheses
           ## (e.g. parasite infections, which would be available as
           ## derived sample variables) should be addressed as part of
           ## the response variable
           jS_loc = Locality
           ) %>%
    full_join(aggregateDensity)


head(allSData)
```


## feed it back into the core data set after dropping the same NA mice as for the sample data

```{r}
PS.TSS_filtered_jS <- subset_samples(PS.TSS_filtered, !is.na(Body_Length))
PS.TSS_filtered_jS@sam_data <- sample_data(allSData)

## check phyloseq object is correct
sample_names(PS.TSS_filtered_jS@otu_table)
sample_names(PS.TSS_filtered_jS@sam_data)

## make sure the sample names are the same for the sample_data and otu_table. We use the jS_mouseID column
sample_names(PS.TSS_filtered_jS@sam_data)<- PS.TSS_filtered_jS@sam_data$jS_mouseID

saveRDS(PS.TSS_filtered_jS, "./intermediate_data/PSvar.rds")
# PS.TSS_filtered_jS <- readRDS("./intermediate_data/PSvar.rds")
```

## get only the selected variables
```{r}

## get selected variables
miceData <- unclass(sample_data(PS.TSS_filtered_jS)) %>% 
  as_tibble() %>% 
  dplyr::select(starts_with("jS_")) 

head(miceData)
nrow(miceData)
# [1] 607 mice
```

## check for missing coordinates in data
```{r}
summary(miceData)

miceData %>% 
  filter(is.na(jS_lon))

head(miceData)
nrow(miceData)
```



##############################
# Extract environmental data #
##############################

```{r prepare workspace}
source("./rscripts/source_pkgs.R")

wrkdir <- getwd()
geoproc_wd <- paste0(wrkdir, "/output/geo_proc")
```


### load data
```{r}
## make mice data spatial 
mice_sf <- st_as_sf(miceData, coords = c("jS_lon", "jS_lat"), crs = 4326, remove = FALSE) %>% 
  st_transform(crs = 3035)
mice_sf

plot(mice_sf["jS_year"])

## german borders
germany <- st_read("./data/geo_raw/germany_3035.gpkg")
german_border <- st_union(germany)

## environmental data
env_cov <- rast(paste0(geoproc_wd, "/stack_env_covariates_germany_100m_3035.tif"))
```

### check data
```{r}
tmap_mode("view")

tm_shape(env_cov$gem_char) +
  tm_raster(legend.show = FALSE) +
  tm_shape(german_border) +
  tm_borders("red") +
  tm_shape(mice_sf) +
  tm_dots("jS_year", size = 1, fill.scale = tm_scale_categorical(values = c("red", "orange", "green", "blue"))) +
  tm_compass() +
  tm_scalebar()
```


```{r}
# reclassify clc table
clc_newclasses <- read.csv(paste0(geoproc_wd, "/clc_reclassification.csv"))
head(clc_newclasses)
unique(clc_newclasses$clc_reclass)
clc_newclasses$CODE_18 <- as.character(clc_newclasses$CODE_18)

## get table of values in the raster
clc2 <- env_cov$clc
plot(clc2)

clc_data <- as.data.frame(cats(clc2))

## add our new columns
clc_new <- clc_data %>% 
  left_join(clc_newclasses, by = c("clc" = "CODE_18"))

## assign values of the new columns to a new raster
clc_reclass_fct <- categories(clc2, layer = 1, value = clc_new, active = 4) # column 7 of data: clc_fct without counting the "Values" column
plot(clc_reclass_fct)

## make a stack
env_cov <- c(env_cov, clc_reclass_fct) 
env_cov
```

## look at data
```{r}
terra::plet(env_cov, "imperv")

plot(env_cov$swf_perc)
plot(mice_sf, add = TRUE, col = "red", pch = 16)
```


# extract values for mice
```{r}
cov_mice <- terra::extract(env_cov, mice_sf)

head(cov_mice)

## put back together
mice_sf <- cbind(mice_sf, cov_mice)

head(mice_sf)

## as data frame
mice_data_env100m <- mice_sf %>% 
  st_drop_geometry() %>% 
  as.data.frame()

head(mice_data_env100m)
colnames(mice_data_env100m)[colnames(mice_data_env100m) == "grids_germany_multi_annual_soil_moist_1991.2020_12"] <- "soil_moist"
colnames(mice_data_env100m)[colnames(mice_data_env100m) == "thuenen_soil_ph_2021"] <- "soil_ph"

write.csv(mice_data_env100m, "./intermediate_data/Table_mice_envcov_100m.csv", row.names = FALSE)
```
